Refactor update_linknets, move from get -> update module
